backend new game done
