+++ app/src/main/java/com/example/firstgame/SplashActivity.java
@@ -1,14 +1,10 @@
 package com.example.firstgame;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.os.CountDownTimer;
-import android.view.View;
-import android.widget.Button;
 
 import com.victor.loading.newton.NewtonCradleLoading;
 
@@ -32,7 +28,7 @@ public class SplashActivity extends Activity {
 
             @Override
             public void onFinish() {
-                Intent intent = new Intent(SplashActivity.this, Menu.class);
+                Intent intent = new Intent(SplashActivity.this, MenuActivity.class);
                 startActivity(intent);
                 finish();
             }
