+++ app/src/main/res/layout/activity_menu.xml
@@ -1,62 +1,67 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#474545"
-    tools:context=".Menu">
+    android:background="@drawable/background_menu"
+    android:orientation="vertical"
+    tools:context=".MenuActivity" >
 
-    <LinearLayout
+    <ImageView
+        android:id="@+id/imageView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:layout_marginBottom="150dp"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        android:src="@drawable/menu_text" />
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
+    <FrameLayout
+        android:id="@+id/layout1"
+        android:clickable="true"
+        android:focusable="true"
+        android:layout_width="match_parent"
+        android:layout_height="100dp"
+        android:layout_margin="15dp"
+        android:background="@drawable/paper">
 
-            <Button
-                android:id="@+id/btn1"
-                android:layout_width="150dp"
-                android:layout_height="150dp"
-                android:layout_weight="1"
-                android:background="#00ACC1"
-                android:text="o'yinga o'tish"
-                android:textSize="18sp" />
+        <ImageView
+            android:id="@+id/imageView1"
+            android:layout_width="288dp"
+            android:layout_height="52dp"
+            android:layout_gravity="center"
+            android:src="@drawable/find_number" />
+    </FrameLayout>
 
-            <Button
-                android:id="@+id/btn2"
-                android:layout_width="150dp"
-                android:layout_height="150dp"
-                android:layout_weight="1"
-                android:background="#F4511E"
-                android:text="topishmoq"
-                android:textSize="18sp" />
-        </LinearLayout>
+    <FrameLayout
+        android:id="@+id/layout2"
+        android:clickable="true"
+        android:focusable="true"
+        android:layout_width="match_parent"
+        android:layout_height="100dp"
+        android:layout_margin="15dp"
+        android:background="@drawable/paper">
 
-        <LinearLayout
-            android:id="@+id/btn3"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="horizontal">
+        <ImageView
+            android:id="@+id/imageView2"
+            android:layout_width="288dp"
+            android:layout_height="52dp"
+            android:layout_gravity="center"
+            android:src="@drawable/question" />
+    </FrameLayout>
 
-            <Button
-                android:id="@+id/btn"
-                android:layout_width="150dp"
-                android:layout_height="137dp"
-                android:layout_weight="1"
-                android:background="#43A047"
-                android:text="gggg"
-                android:textSize="18sp" />
+    <FrameLayout
+        android:id="@+id/layout3"
+        android:clickable="true"
+        android:focusable="true"
+        android:layout_width="match_parent"
+        android:layout_height="100dp"
+        android:layout_margin="15dp"
+        android:background="@drawable/paper">
 
-        </LinearLayout>
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
+        <ImageView
+            android:id="@+id/imageView3"
+            android:layout_width="288dp"
+            android:layout_height="52dp"
+            android:layout_gravity="center"
+            android:src="@drawable/gallery" />
+    </FrameLayout>
+</LinearLayout>
